Extract TrackList from Editor.tsx (verbatim)
Move the ~760-line tracks section (and its collapse / naddr-resolution
state + handlers) into AlbumEditor/TrackList.tsx, a verbatim mechanical
move with no behavior change. Editor.tsx drops from ~1080 to ~320 lines
and is now a thin composition of the AlbumEditor section components.
isEnabled is threaded in (FeatureId-typed) for the per-track value gate.

Build + lint + 104 tests pass.
